backend/requirements.txt
fastapi
uvicorn[standard]
redis
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
pymssql
pymysql
httpx
requests
